% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{boxed_theme}
\alias{boxed_theme}
\title{A boxed theme for ggplot2}
\usage{
boxed_theme(size_lg = 6, size_sm = 5)
}
\arguments{
\item{size_lg}{size of axis titles}

\item{size_sm}{size of all other text}
}
\description{
A boxed theme for ggplot2
}
