% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{clean_theme}
\alias{clean_theme}
\title{A clean theme for ggplot2}
\usage{
clean_theme(size_lg = 6, size_sm = 5)
}
\arguments{
\item{size_lg}{size of axis titles}

\item{size_sm}{size of all other text}
}
\description{
A clean theme for ggplot2
}
