% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{darken}
\alias{darken}
\title{Darken a color}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{color}{the color to darken}

\item{factor}{amount of darkening to apply}
}
\description{
Darken a color
}
