% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_system.R
\name{detect_system}
\alias{detect_system}
\title{Detect the system we are working on}
\usage{
detect_system(project_name = NULL)
}
\arguments{
\item{project_name}{optionally, specify the name of the current project.
If unset, will be inferred from the working directory}
}
\value{
assigns values to global variables \code{system} and \code{base_dir}
}
\description{
Detect the current system (e.g., sockeye, LSI cluster, cedar) and set the
base directory accordingly.
}
