% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Create scientific labels for ggplot}
\usage{
fancy_scientific(l)
}
\arguments{
\item{l}{values to transform}
}
\description{
Create scientific labels for ggplot
}
