% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{lighten}
\alias{lighten}
\title{Lighten a color}
\usage{
lighten(color, factor = 1.4)
}
\arguments{
\item{color}{the color to lighten}

\item{factor}{amount of lightening to apply}
}
\description{
Lighten a color
}
