% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{umap_theme}
\alias{umap_theme}
\title{A UMAP theme for ggplot2}
\usage{
umap_theme(size_lg = 6, size_sm = 5)
}
\arguments{
\item{size_lg}{size of axis titles}

\item{size_sm}{size of all other text}
}
\description{
A UMAP theme for ggplot2
}
