% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{submit_job}
\alias{submit_job}
\title{Submit a job to the scheduler}
\usage{
submit_job(grid, script, allocation, job_loop = 1)
}
\arguments{
\item{grid}{path to the tab-delimited grid file}

\item{script}{path to write the submission script to}

\item{allocation}{account name to use on the platform in question}

\item{job_loop}{how many rows in the grid file should be run in one array
job?}
}
\description{
Submit an array of jobs to the scheduler on a high-performance computing 
cluster.
}
