% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sh.R
\name{write_sh}
\alias{write_sh}
\title{Write job submission script (slurm or PBSPro)}
\usage{
write_sh(
  job_name,
  sh_file,
  grid_file,
  inner_file,
  env = "decoy-generation/env",
  time = 24,
  mem = 4,
  cpus = 1,
  gpu = FALSE
)
}
\arguments{
\item{job_name}{name of the job that will appear in the job management system}

\item{sh_file}{path to write the submission script to}

\item{grid_file}{path to the tab-delimited grid file}

\item{inner_file}{path to the R/python inner file}

\item{env}{path to the conda environment}

\item{time}{time of the job, in hours}

\item{mem}{memory required for the job, in GB}

\item{cpus}{CPUs required for the job}

\item{gpu}{if \code{TRUE}, a GPU will be requested}
}
\value{
returns nothing, but writes the slurm/PBSPro script to the specified
  filepath
}
\description{
Write a submission script to run an array of jobs on a high-performance 
computing platform. This function abstracts away specifying the computational
requirements of the job, loading the relevant conda environment, and then
reading the arguments from the grid file and passing them into an R or python
'inner' file.
}
